+++ src/main/java/com/fasipe/biomedicina/Requests/RequestProcedimentos.java
@@ -3,7 +3,7 @@ package com.fasipe.biomedicina.Requests;
 public record RequestProcedimentos(
 
         Long id,
-        Integer codProced,
+        Long codProced,
 
         String descProced,
